+++ docs/src/.vitepress/components/Button.vue
@@ -1,8 +1,0 @@
-<template>
-  <iframe src="https://codesandbox.io/embed/determined-voice-udq8lt?fontsize=14&hidenavigation=1&theme=dark&view=preview"
-     style="width:100%; height:300px; border:0; border-radius: 6px; overflow:hidden;"
-     title="determined-voice-udq8lt"
-     allow="accelerometer; ambient-light-sensor; camera; encrypted-media; geolocation; gyroscope; hid; microphone; midi; payment; usb; vr; xr-spatial-tracking"
-     sandbox="allow-forms allow-modals allow-popups allow-presentation allow-same-origin allow-scripts"
-   />
-</template>