+++ packages/react/src/components/IconButton/IconButton.stories.tsx
@@ -34,7 +34,7 @@ export default {
       }
     },
     size: {
-      options: ['sm', 'md', 'lg'],
+      options: ['xs', 'sm', 'md', 'lg'],
       control: {
         type: 'inline-radio'
       }